Thresholder
note: NO constructor without arguments
