+++ src/PqExtra.h
@@ -29,54 +29,4 @@
 #include "MovingStats.h"
 #include "SimpleStats.h"
 
-#define THRESHOLD_LOW     0
-#define THRESHOLD_HIGH    1
-#define THRESHOLD_FALLING 2
-#define THRESHOLD_RISING  3
-#define THRESHOLD_CHANGE  4
-
-/**
- * Emits a signals that tells if a certain threshold has been crossed.
- * Possible modes are:
- * - THRESHOLD_LOW  : true iff value < threshold
- * - THRESHOLD_HIGH : true iff value > threshold
- * - THRESHOLD_FALLING : true when value becomes < threshold, then waits until it becomes > resetThreshold (*)
- * - THRESHOLD_RISING  : true when value becomes > threshold, then waits until it becomes < resetThreshold (*)
- * - THRESHOLD_CHANGE  : true when value crosses threshold
- * (*) If resetThreshold is not specified, it is set to threshold by default. Parameter resetThreshold is only
- * used in the FALLING and RISING modes, otherwise it is ignored.
- */
-class Thresholder : public PqPutter, public PqDigitalGetter {
-public:
-  Thresholder(float threshold, uint8_t mode=THRESHOLD_HIGH);
-  Thresholder(float threshold, uint8_t mode, float resetThreshold);
-  virtual ~Thresholder() {}
-
-  /**
-   * Pushes value into the unit.
-   * @param value the value sent to the unit
-   * @return the new value of the unit
-   */
-  virtual float put(float value);
-
-  /// Returns true iff the threshold is crossed.
-  virtual bool isOn() { return _value; }
-
-protected:
-	// Threshold values.
-  float _threshold;
-	float _resetThreshold;
-
-	// Thresholding mode.
-  uint8_t  _mode  : 3;
-
-	// Booleans used to keep track of signal value.
-  bool _value     : 1;
-	bool _wasLow    : 1;
-	bool _wasHigh   : 1;
-
-private:
-	void _init(float threshold, uint8_t mode, float resetThreshold);
-};
-
 #endif
